requirements.txt
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1600699246375/work
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imageio==2.9.0
kiwisolver==1.2.0
matplotlib @ file:///C:/ci/matplotlib-base_1597876438601/work
mkl-fft==1.2.0
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.5
numpy @ file:///C:/ci/numpy_and_numpy_base_1596215850360/work
olefile==0.46
opencv-python==4.4.0.44
pandas @ file:///C:/ci/pandas_1598371563714/work
Pillow @ file:///C:/ci/pillow_1594298230227/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyparsing==2.4.7
python-dateutil==2.8.1
PyTurboJPEG @ git+git://github.com/lilohuang/PyTurboJPEG.git@614e1edd155f468ecce10ccfa8efba82d960d1ae
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy @ file:///C:/ci/scipy_1592916963468/work
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
sip==4.19.13
six==1.15.0
tifffile==2020.9.30
torch @ file:///C:/ci/pytorch_1596373105144/work
torchvision==0.7.0
tornado==6.0.4
wincertstore==0.2
